Add middleware to validate whether user is logged or not
