+++ src/pages/cinema/_middleware.ts
@@ -1,0 +1,8 @@
+import { withAuth } from 'next-auth/middleware'
+
+export default withAuth({
+  pages: {
+    signIn: '/',
+    error: '/'
+  }
+})
