+++ src/pages/index.tsx
@@ -1,28 +1,36 @@
-import { Button, Flex } from "@chakra-ui/react";
-import { Sidebar } from "../components/Sidebar";
-import { getSession, signOut } from 'next-auth/react'
-import { GetServerSideProps } from "next";
+import { Button, Flex, Heading } from "@chakra-ui/react";
+import { GetServerSideProps, NextPage } from "next";
+import { getSession, signIn, useSession } from "next-auth/react";
+import { FaGoogle } from 'react-icons/fa'
 
-export default function Dashboard() {
+const Login: NextPage = () => {
+  const { data: session  } = useSession()
+
+  const handleSignInWithGoogle = async () => {
+    if(!session) {
+      await signIn('google')
+    }
+  }
+
   return (
-    <Flex direction="column" h="100vh">
-      <Flex
-        w="100vw"
-      >
-        <Sidebar />
-
+    <Flex w="100vw" h="100vh" align="center" justify="center">
+      <Flex flexDir="column" maxW="90vw" borderWidth={1} borderColor="gray.800" borderRadius="5" p="2rem" w="500px">
+          <Heading as="h2" fontSize="1.6rem" mb="2rem">Laboratório Maker</Heading>
+            <Button leftIcon={<FaGoogle color="white" />} flex="1" w="100%" onClick={handleSignInWithGoogle}>Acessar com Google</Button>
       </Flex>
     </Flex>
   )
 }
 
+export default Login;
+
 export const getServerSideProps: GetServerSideProps = async ({ req, params }) => {
   const session = await getSession({ req })
 
-  if(!session) {
+  if(!!session) {
     return {
       redirect: {
-        destination: '/login',
+        destination: '/dashboard',
         permanent: false,
       }
     }
